backend/ai/face_guard/requirements.txt
ultralytics
opencv-python
numpy
insightface
faiss-cpu
onnxruntime
